refactor: MyPageService의 서비스 로직을 StudentService로 이동시키고 MyPageService 클래스 삭제 (#82)
